requirements.txt
fastapi
uvicorn
pydantic
mistralai
qdrant-client
python-dotenv
requests
reportlab
